Update merchantSafe, profilescreen refactor
