+++ cardstack/src/screens/ProfileScreen/components/index.ts
@@ -1,0 +1,2 @@
+export * from './fixture';
+export * from './CreateProfileStepOne';
